第二套完成

+++ phonehelper-second/app/src/main/java/com/mob/sms2/service/UpdateService.java
@@ -58,7 +58,7 @@ public class UpdateService extends IntentService {
                         if (data != null) {
                             String versioncode = data.getNo();
                             try {
-                                int code = Integer.parseInt(data.getNo());
+                                double code = Double.parseDouble(data.getNo());
                                 if (BuildConfig.VERSION_CODE < code) {
                                     // 升级
                                     WeakReference<Activity> currentActivity = AppManager.getCurrentActivity();
